src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==24.1.0
langchain==0.2.6
PyMuPDF==1.24.7
langchain-community==0.2.6
motor==3.5.0